requirements.txt
aniso8601==8.1.1
anyio==4.9.0
blockchain==1.4.4
certifi==2020.4.5.1
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click>=8.0.0
cryptocompare==0.7.0
cryptography==3.3.2
decorator==4.4.2
dnspython==2.7.0
emoji==0.5.4
enum-compat==0.0.3
Flask==2.2.5
Flask-RESTful>=0.3.9
future==0.18.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==2.9
itsdangerous>=2.0
Jinja2>=3.0
MarkupSafe>=2.0
pycparser==2.20
pycryptodome==3.9.7
pymongo==4.3.3
pyTelegramBotAPI==3.6.7
python-dotenv==0.13.0
python-telegram-bot==22.0
pytz==2021.1
Quart==0.18.4
requests==2.32.3
six==1.14.0
sniffio==1.3.1
SQLAlchemy==1.3.16
tornado==6.0.4
urllib3==2.3.0
Werkzeug==2.2.3
